- refactor activators - organize packages for now   - could do better
