+++ bootlaces/src/main/java/com/candroid/bootlaces/activators/BackgroundActivatorReceiver.kt
@@ -12,12 +12,15 @@ distributed under the License is distributed on an "AS IS" BASIS,
 WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 See the License for the specific language governing permissions and
 limitations under the License.*/
-package com.candroid.bootlaces
+package com.candroid.bootlaces.activators
 
 import android.content.BroadcastReceiver
 import android.content.Context
 import android.content.Intent
 import android.os.Build
+import com.candroid.bootlaces.service.BootServiceState
+import com.candroid.bootlaces.service.notification.Boot
+import com.candroid.bootlaces.service.notification.IBoot
 import dagger.hilt.android.AndroidEntryPoint
 import kotlinx.coroutines.ExperimentalCoroutinesApi
 import javax.inject.Inject
@@ -50,7 +53,7 @@ import javax.inject.Inject
  * Starts BootService after the phone turns on.
  **/
 @AndroidEntryPoint
-class BootReceiver : HiltBugReceiver<Boot>(){
+class BackgroundActivatorReceiver : HiltBugReceiver<Boot>(){
    @Inject lateinit var boot: IBoot
 
     @ExperimentalCoroutinesApi
