+++ bootlaces/src/main/java/com/candroid/bootlaces/api/ForegroundService.kt
@@ -18,11 +18,9 @@ import android.app.Notification
 import androidx.datastore.DataStore
 import androidx.datastore.preferences.Preferences
 import androidx.lifecycle.LifecycleCoroutineScope
-import com.candroid.bootlaces.IBoot
+import com.candroid.bootlaces.service.notification.IBoot
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.flow.Flow
-import kotlinx.coroutines.flow.flowOn
-import kotlinx.coroutines.flow.map
 
 /*
             (   (                ) (             (     (
